refactor(keyboard-functions): abstract action handler
